backend/requirements.txt
flask
flask-cors
numpy
opencv-python-headless
pycryptodome
gunicorn
